Changed LONG_MAX to INT_MAX
git-svn-id: svn://orca.physics.unc.edu/Orca64@11945 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Custom Hardware/SNO+/hiredis/dict.c
@@ -319,7 +319,7 @@ static int _dictExpandIfNeeded(dict *ht) {
 static uint32_t _dictNextPower(uint32_t size) {
     uint32_t i = DICT_HT_INITIAL_SIZE;
 
-    if (size >= LONG_MAX) return LONG_MAX;
+    if (size >= INT_MAX) return INT_MAX;
     while(1) {
         if (i >= size)
             return i;
